arreglo de migraciones

+++ database/migrations/2020_03_18_150706_create_multas_recargas_table.php
@@ -17,7 +17,7 @@ class CreateMultasRecargasTable extends Migration
             $table->bigIncrements('id');
             $table->string('motivo');
             $table->text('observacion')->nullable();
-            $table->float('monto');
+            $table->double('monto');
             $table->enum('tipo',['Multa','Recarga']);
             $table->integer('anio');
             $table->unsignedBigInteger('id_admin');
